Create Book component for rendering single book
