+++ src/components/BookDB.jsx
@@ -1,0 +1,22 @@
+const primaryBooks = [
+  {
+    title: 'The Lord of the Rings',
+    author: 'J R R Tolkien',
+    category: 'Fantasy',
+    id: 1,
+  },
+  {
+    title: 'Deception Point',
+    author: 'Dan Brown',
+    category: 'Thriller',
+    id: 2,
+  },
+  {
+    title: 'Hamlet',
+    author: 'Shakespeare',
+    category: 'Classic',
+    id: 3,
+  },
+];
+
+export default primaryBooks;
