+++ src/components/Booklist.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import Book from './Book';
 
 const Booklist = ({ books }) => {
   const [bookList, setBookList] = useState(books);
@@ -20,17 +21,7 @@ const Booklist = ({ books }) => {
       <ul>
         {bookList.map((book) => (
           <li key={book.id}>
-            <span>
-              <strong>Title:</strong>
-              {book.title}
-              ,
-              <strong>Author:</strong>
-              {book.author}
-              ,
-              <strong>Category:</strong>
-              {book.category}
-            </span>
-            <button type="button" onClick={() => handleDeleteBook(book.id)}>Delete</button>
+            <Book book={book} deleteBook={handleDeleteBook} />
           </li>
         ))}
       </ul>
